run haproxy with docker in tests.
Extend tox configuration so that haproxy is automatically started with
docker in order to run the haproxy control tests with tox.

Also configure travis so that we can use docker.

This adds Docker as testing dependency but makes testing more convenient
and now all components are included in tox / travis tests.