Skip events when HaProxy connection refused.
When the HaProxy stats socket is not reachable and results in a refused
connection, events can pile up and lead to a high load of the supervisor
and the eventlistener processes.

Reasons for this situation could be that HaProxy is not running at all,
the stats socket is configured inproperly, or the event listener is not
configured correctly (e.g. not on the same server).

In order to avoid generating high load a threshold is introduced: only
the first 4 consecutive refused connections are responeded with a FAIL
to supervisor (which requeues the event), then the eventlistener starts
dropping events and responding with OK without even trying to send them
to the HaProxy stats socket.

After a timeout of one minute, where everything is dropped, the
eventlistener will retry to send events.

With this approach we are no longer generating high load when the
HaProxy is never reachable, but when a single connection cannot be
established because of a temporary problem (e.g. network instability),
the event requeued and retried.